LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/015-expected.txt


FAIL  Link with javascript onclick and href script order  assert_array_equals: expected property 1 to be "after script" but got "href" (expected array ["click", "after script", "href"] got ["click", "href", "after script"])
 Test
